continue applying scenarios
