+++ requirements-dev.txt
@@ -1,4 +1,5 @@
 pytest
 pytest-asyncio
 pytest-cov
+pytest-xdist
 python-kraken-sdk>=3.2.6,<4.0.0
